Create an interactive webpage to play a game of RPS
This commit adds javaScript code to create a game of rock, paper,
scissors in a webpage.

It allows the player to make moves and keep track of the score.

The HTML was updated and CSS classes added to create a responsive
design.

A play game function was added that includes DOM manipulation. It also
includes:

A game end when a player reaches a score of 5. The RPS buttons then
disable and a the round result message changes to an end game message.

A reset button that can reset the player scores to 0. This can be done
at any point during the game.